Refactor structure
-  `base` contains base functionality potentially useful for different action impls
-  `devTo` contains the implementation of the `devTo` command

JIRA: RAIL-1815
