+++ tools/applink/src/base/dependencyDiscovery.ts
