+++ tools/applink/src/base/utils.ts
